패키지 네이밍 변경, member 관련 내용추가
