+++ project/src/main/java/com/campers/camfp/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.campers.camfp;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(ProjectApplication.class);
+	}
+
+}
