+++ project/src/main/java/com/campers/camfp/entity/Board.java
@@ -1,0 +1,46 @@
+package com.campers.camfp.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@Table(name="board")
+@ToString
+@Builder
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+public class Board {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long bno;
+	
+	@Column(nullable = false)
+	private String title;
+	
+	@Column(length = 2000, nullable = false)
+	private String content;
+	
+	@ManyToOne(fetch = FetchType.LAZY)
+	private User user;
+	
+	private String category;
+	
+	private int count;
+	
+	private int heart;
+	
+}
