+++ project/src/main/java/com/campers/camfp/entity/Camp.java
@@ -1,0 +1,33 @@
+package com.campers.camfp.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@ToString
+public class Camp {
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long cno;
+	
+	private String name;
+	private String country;
+	private String address;
+	private int unit;
+	private int count;
+	private int star;
+
+}
