+++ project/src/main/java/com/campers/camfp/entity/Reply.java
@@ -1,0 +1,42 @@
+package com.campers.camfp.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Getter
+@Table(name="reply")
+@ToString(exclude = {"board", "user"})
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Reply {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long rno;
+	
+	@ManyToOne(fetch = FetchType.LAZY)
+	private Board board;
+	
+	@Column(length = 2000, nullable = false)
+	private String content;
+	
+	@ManyToOne(fetch = FetchType.LAZY)
+	private User user;
+	
+	private int heart;
+	
+}
