+++ project/src/main/java/com/campers/camfp/repository/CampRepository.java
@@ -1,0 +1,9 @@
+package com.campers.camfp.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.campers.camfp.entity.Camp;
+
+public interface CampRepository extends JpaRepository<Camp, Long> {
+
+}
