--- project/src/main/java/com/emptysky/project/ProjectApplication.java
@@ -1,15 +1,0 @@
-package com.emptysky.project;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class ProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProjectApplication.class, args);
-	}
-
-}
