--- project/src/main/java/com/emptysky/project/dto/ItemDTO.java
@@ -1,30 +1,0 @@
-package com.emptysky.project.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ItemDTO {
-
-	private Long ino;
-	private String name;
-	private String image;
-	private String brand;
-	private String maker;
-	private String category1;
-	private String category2;
-	private String category3;
-	private int price;
-	private String link;
-	private int type;
-	private int count;
-	private int star;
-	private int heart;
-	
-	
-}
