--- project/src/main/java/com/emptysky/project/entity/ShoppingCart.java
@@ -1,43 +1,0 @@
-package com.emptysky.project.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import org.springframework.beans.factory.annotation.Value;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Table(name="shopping_cart")
-@ToString(exclude = {"user", "item"})
-@Getter
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ShoppingCart {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long sno;
-	
-	@Value("1")
-	private int amount;
-	
-	@ManyToOne(fetch = FetchType.LAZY)
-	private User user;
-	
-	@ManyToOne(fetch = FetchType.LAZY)
-	private Item item;
-	
-	
-	
-}
