--- project/src/main/java/com/emptysky/project/repository/ItemRepository.java
@@ -1,9 +1,0 @@
-package com.emptysky.project.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.emptysky.project.entity.Item;
-
-public interface ItemRepository extends JpaRepository<Item, Long>{
-
-}
