+++ project/src/main/resources/templates/sample/login.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+<meta charset="UTF-8">
+<title>Login Page</title>
+</head>
+<body>
+	<h1>이곳은 로그인 페이지입니다.</h1>
+</body>
+</html>
