+++ project/src/test/java/com/campers/camfp/MemberTestApplicationTests.java
@@ -1,0 +1,13 @@
+package com.campers.camfp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MemberTestApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
