--- project/src/test/java/com/emptysky/project/ProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.emptysky.project;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
