--- project/src/test/java/com/emptysky/project/repository/CampRepositoryTests.java
@@ -1,32 +1,0 @@
-package com.emptysky.project.repository;
-
-import java.util.stream.IntStream;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import com.emptysky.project.entity.Camp;
-
-@SpringBootTest
-public class CampRepositoryTests {
-
-	@Autowired
-	private CampRepository campRepository;
-	
-	@Test
-	public void insertCamp() {
-		
-		IntStream.rangeClosed(0, 100).forEach(i -> {
-			
-			Camp camp = Camp.builder()
-							.count((int)(Math.random()*100) + 1)
-							.address("한국 " + i + "번지")
-							.country("대한민국 " + i + "도")
-							.build();
-			campRepository.save(camp);
-			
-		});
-	}
-	
-}
